[DAEF-232] finishes up the integration of numeric input
