+++ package.json
@@ -6,7 +6,7 @@
   "main": "./dist/main.js",
   "scripts": {
     "lint": "eslint --format=node_modules/eslint-formatter-pretty app features *.js",
-    "hot-server": "cross-env NODE_ENV=development node -r babel-register webpack/server.js",
+    "hot-server": "cross-env NODE_ENV=development node --preserve-symlinks -r babel-register webpack/server.js",
     "build-main": "cross-env NODE_ENV=production node -r babel-register ./node_modules/webpack/bin/webpack --config webpack/webpack.config.electron.js --progress --profile --colors",
     "build-renderer": "cross-env NODE_ENV=production node -r babel-register ./node_modules/webpack/bin/webpack --config webpack/webpack.config.production.js --progress --profile --colors",
     "build": "npm run build-main && npm run build-renderer",
@@ -140,7 +140,7 @@
     "react-intl": "2.2.3",
     "react-markdown": "2.5.0",
     "react-number-format": "1.1.2",
-    "react-polymorph": "0.2.4",
+    "react-polymorph": "0.2.6",
     "react-router": "3.0.3",
     "react-toolbox": "1.3.4",
     "recharts": "0.19.1",
